[STYLE] 컴포넌트화
