+++ src/shared/Router.js
@@ -15,7 +15,7 @@ const Router = () => {
         <Route path="/common/classes" element={PAGE.member} />
         <Route path="/common/classes/:id" element={PAGE.member} />
         <Route
-          path="/host/:classId/:scheduleId?/:timeId?"
+          path="/host/:classroomId/:scheduleId?/:timeId?"
           element={PAGE.host}
         />
         <Route path="/preview" element={PAGE.preview} />
